% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brspplot.R
\name{brsp_subregions_plot}
\alias{brsp_subregions_plot}
\title{brsp_subregions_plot}
\usage{
brsp_subregions_plot(gene, technique = "array", cmp = heat.colors,
  log_transform = FALSE, im_height = 25, im_width = 75, save_pdf = TRUE)
}
\arguments{
\item{gene}{Gene in the `brspdata` package}

\item{technique}{Either `rna`, for RNAseq, or `array`, for MicroArray}

\item{cmp}{R color-map function for ages; default heat.colors}

\item{log_transform}{Boolean for plotting of linear or log data}

\item{im_height}{PDF image height}

\item{im_width}{PDF image width}

\item{save_pdf}{Save PDF or output to console}
}
\value{
Creates a new plot of the gene in the current working directory
}
\description{
Plot subregion expression levels for the human brainspan atlas. Provides
coloring by both age and ontology in a single plot.
}
\examples{
# call on a given gene
brsp_subregions_plot("SCARF1", "rna")
}
