% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\name{clear_cache}
\alias{clear_cache}
\title{clear_cache}
\usage{
clear_cache()
}
\description{
Deletes all current files in the cache. This is implemented in the case that an item
on the API updates. Clearing the cache will make it so the new item is downloaded at
next run. Not necessary for short term use.
}
\examples{
# delete cache items since a new ontology has been released!
clear_cache()

}
