% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetplot.R
\name{fet_brain_location_expression_plot}
\alias{fet_brain_location_expression_plot}
\title{fet_brain_location_expression_plot}
\usage{
fet_brain_location_expression_plot(gene)
}
\arguments{
\item{gene}{Gene in the `fetdata::datFET` data set; character string}
}
\value{
Creates a new plot of the gene in the current working directory; label as `GENE_fetalHuman_BrainLocationExpressionPlotsForEachLayer.pdf`
}
\description{
Creates a matrix of brain images. Each column corresponds to a brain in the fetal human
brain atlas, while each row is a specific brain layer.  Each point shows gene expression
for a particular cortical region in it's approximate anatomical centroid in the background
brain image.  Requires data from https://github.com/AllenInstitute/fetdata.
}
\examples{
# call on a given gene
fet_expression2D_plot("SHH")

}
