% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetplot.R
\name{fet_subregions_plot}
\alias{fet_subregions_plot}
\title{fet_subregions_plot}
\usage{
fet_subregions_plot(gene, log_transform = FALSE, im_height = 10,
  im_width = 50, save_pdf = TRUE)
}
\arguments{
\item{gene}{Gene in the `fetdata::datFET` data set; character string}

\item{log_transform}{Boolean for plotting of linear or log data}

\item{im_height}{PDF image height}

\item{im_width}{PDF image width}

\item{save_pdf}{Save to disk or output to console}
}
\value{
Creates a new plot of the gene in the current working directory; label as `GENE_fetalHuman_structureBarPlotAll.pdf`
}
\description{
Plot gene expression values for each structure in each donor for the fetal human brain atlas. This function
requires the fetdata package (GITHUB GOES HERE), and is a thin wrapper
around a modified version of WGCNA's verboseBarplot
}
\examples{
# call on a given gene
fet_subregions_plot("SHH")
}
