% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbaplot.R
\name{hba_subregions_plot}
\alias{hba_subregions_plot}
\title{hba_subregions_plot}
\usage{
hba_subregions_plot(
  gene,
  log_transform = FALSE,
  im_height = 20,
  im_width = 32,
  save_pdf = TRUE
)
}
\arguments{
\item{gene}{Gene in the `hbadata::datHBA` data set; character string}

\item{log_transform}{Boolean for plotting of linear or log data}

\item{im_height}{PDF image height}

\item{im_width}{PDF image width}

\item{save_pdf}{Output pdf or plot object}
}
\value{
Creates a new plot of the gene in the current working directory; label as `GENE_HBA_subregionsPlot.pdf`
}
\description{
Plot gene expression values for each structure in each donor. This function
requires the hbadata package (https://github.com/AllenInstitute/hbadata), and
is a thin wrapper around a modified version of WGCNA's verboseBarplot
}
\examples{
# call on a given gene
hba_subregions_plot("FOXP2")
}
