% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_plots.R
\name{mouse_structureplot}
\alias{mouse_structureplot}
\title{mouse_structureplot}
\usage{
mouse_structureplot(gene, atlas, experiment = NULL, struct_depth = 3)
}
\arguments{
\item{gene}{Gene in the Allen Mouse or Developing Mouse atlas; character string}

\item{atlas}{Atlas to plot the gene for; either "DevMouse" or "Mouse"}

\item{experiment}{Experiment ID; One gene may have multiple experiments. If no values is supplied a list of valid experiments will be presented to the user}

\item{struct_depth}{ontology structure depth; default of 3}
}
\value{
Creates a new plot of the gene in the current working directory
}
\description{
Plot gene expression values for strucutures at a given ontology depth for
the mouse or developing mouse atlas. This function pull the mean expression
energies directly from the Allen institute API. Utilizes WGCNAs 
verboseBarplot function for visualization.
}
\examples{
# call on a given gene
mouse_structureplot("Scarf1", atlas="Mouse")

# call a given gene with a specific experiment
mouse_structureplot("Shh", atlas="DevMouse", experiment="100093339")
}
