% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_structure_plots.R
\name{mouse_subregions_plot}
\alias{mouse_subregions_plot}
\title{mouse_subreginos_plot}
\usage{
mouse_subregions_plot(gene, atlas, experiment = NULL, struct_depth = 3,
  im_width = NULL, im_height = NULL, save_pdf = TRUE)
}
\arguments{
\item{gene}{Gene in the Allen Mouse or Developing Mouse atlas; character string}

\item{atlas}{Atlas to plot the gene for; either "DevMouse" or "Mouse"}

\item{experiment}{Experiment ID string; Experiment ID for the adult mouse atlas. The devmouse uses all experiments, so the ID is ignored. If not ID is provided, the program will ask you to choose one from a list.}

\item{struct_depth}{ontology structure depth; default of 3}

\item{im_width}{Output image width; optional}

\item{im_height}{Output image height; optional}

\item{save_pdf}{Save plot to disk as pdf or return to console}
}
\value{
Creates a new plot of the gene in the current working directory
}
\description{
Plot gene expression values for strucutures at a given ontology depth for
the mouse or developing mouse atlas. This function pull the mean expression
energies directly from the Allen institute API. Utilizes WGCNAs 
verboseBarplot function for visualization.
}
\examples{
# call on a given gene
mouse_structureplot("Scarf1", atlas="Mouse")

# call a given gene with a specific experiment
mouse_structureplot("Shh", atlas="DevMouse", struct_depth = 4, im_width = 25)
}
