% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpplot.R
\name{nhp_cortex_series_plot}
\alias{nhp_cortex_series_plot}
\title{nhp_cortex_series_plot}
\usage{
nhp_cortex_series_plot(
  gene,
  log_transform = FALSE,
  im_height = 12,
  im_width = 4,
  save_pdf = TRUE
)
}
\arguments{
\item{gene}{Gene in the `nhpdata::exprl2` data set; character string}

\item{im_height}{PDF image height}

\item{im_width}{PDF image width}

\item{save_pdf}{Save to disk or return to console}
}
\value{
Creates a new plot of the gene in the current working directory
}
\description{
Plot times series for a single gene in the non-human primate brain atlas from the Allen
institute. This function relies on the nhpdata package, located at
https://github.com/AllenInstitute/nhpdata.
}
\examples{
# call on a given gene
nhp_cortex_series_plot("MKI67")

}
