% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpplot.R
\name{nhp_subregions_plot}
\alias{nhp_subregions_plot}
\title{nhp_subregions_plot}
\usage{
nhp_subregions_plot(gene)
}
\arguments{
\item{gene}{Gene in the `nhpdata::datnhp` data set; character string}
}
\value{
Creates a new plot of the gene in the current working directory; label as `GENE_nhpalHuman_structureBarPlotAll.pdf`
}
\description{
Plot gene expression values for each structure in each donor for the nhpal human brain atlas. This function
requires the nhpdata package (GITHUB GOES HERE), and is a thin wrapper
around a modified version of WGCNA's verboseBarplot
}
\examples{
# call on a given gene
nhp_subregions_plot("SHH")
}
