% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_system.R
\name{set_cache_size}
\alias{set_cache_size}
\title{set_cache_size}
\usage{
set_cache_size(size)
}
\arguments{
\item{size}{number of bytes allocated to the cache; default 2e9. size = FALSE disables cacheing}
}
\description{
set_cache_size is a atlasplot helper function that allows the user to decide the amount
of storage they'd like to allocate to the on disk file cache. All of the functions
using the Allen Institute API make use of the cache. If cache size is exceeded, the
oldest item are removed.
}
\examples{
# set cache to 10 kilobyte
set_cache_size(10000)

# turn off cache
set_cache_size(FALSE)

}
