% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpplot.R
\name{species_expression_series_plot}
\alias{species_expression_series_plot}
\title{species_expression_series_plot}
\usage{
species_expression_series_plot(
  Gene,
  col_map = rainbow,
  im_width = 10,
  im_height = 10,
  save_pdf = TRUE
)
}
\arguments{
\item{im_width}{PDF image width}

\item{im_height}{PDF image height}

\item{save_pdf}{Save to disk or return to console}

\item{gene}{Gene in the `nhpdata::dev.expr2` data set; character string}

\item{colormap}{Color pallete used to choose plot colors; default rainbow}
}
\value{
Creates a new plot of the gene in the current working directory
}
\description{
Plot gene expression values for a time series of ages for human, non-human primate, mouse,
and rat. Will only include data that is present in plots. This function requires the
nhpdata package https://github.com/AllenInstitute/nhpdata
}
\examples{
# call on a given gene
species_expression_series_plot("EMX2")

# call on new gene with a different color map
species_expression_series_plot("DAD1", col_map = heat.colors))

}
