% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpplot.R
\name{nhp_cortex_expression2D_small_plot}
\alias{nhp_cortex_expression2D_small_plot}
\title{nhp_cortex_expression2D_small_plot}
\usage{
nhp_cortex_expression2D_small_plot(gene, colVec = c("white", "red"))
}
\arguments{
\item{gene}{Gene in the `nhpdata::datNHP` data set; character string}

\item{colVec}{List of colors to create linear color map; c("white", "red") default}
}
\value{
Creates a new plot of the gene in the current working directory
}
\description{
Creates a boxed heatmap of the expression values for a single gene in the non-human
primate atlas. This plot is limited to only the V1 region.
}
\examples{
# call on a given gene
nhp_cortex_expression2D_plot("PAX6", c("green", "skyblue"))

}
